requirements.txt
python-dotenv==1.0.1
openai==1.54.3
requests==2.32.3
telebot==0.0.5